Turn `constraint_system` into a public `r1cs` module
This lets us cut down on the amount of facade to maintain.
